requirements.txt
scipy>=1.2.1
pygmo>=2.10
numpy>=1.16.4
matplotlib>=2.2.2
gym>=0.12.5
python>=3.6.8
PyYAML>=5.3
pickle
jupyter5.3